# Add isascii to normalize_identifier for 25% reduction in parse time