@c42f 2023-03-23
Nice!!  I assume you refer to JuliaLang/julia#48568

Does this also help with the older `isascii` in Julia 1.9?