@ndinsmore 2023-03-24
The old isascii could only do about 2 bytes a cycle in <1.10, The new version will do 20+bytes/cycle.   The UTF8 code was slow so I am assuming the old version should still offer a good benefit.