@KristofferC 2023-03-24
> The old isascii could only do about 2 bytes a cycle in <1.10, The new version will do 20+bytes/cycle. 

Most identifiers are quite short though, and the chunk size in that PR is 1024, so I am not sure you will get out the full speed from it on your average Julia identifier.