@KristofferC 2023-03-24
I totally agree that this thing is the right thing to do. I am just not sure that it is due to the improvements of `isascii` on master.