@c42f 2023-04-11
Julia's reference parser already uses an equivalent optimization internally (see the C code in - `fl_accum_julia_symbol`)